refactor: move layer preferences to repo class
